USB: Update USB probe code and rewrite/reorganize the HID driver.
The previous code quit searching for a driver when a probe failure
used the throw mechanism to exit. This update to the USB probe code
allows the search for a driver to continue after the throw mechanism
is used. This change also required changing all of the calls of (throw
:probe-failed NIL) to (throw :probe-failed :failed) so that the throw
path can be distinguished from the normal failure exit of returning
NIL.

The HID driver was restructured to move the HID mouse driver to
hid-mouse.lisp and the HID keyboard driver stub to
hid-keyboard.lisp. (hid-keyboard.lisp contains a stub driver which
contains no real functionality.)

In addition, the HID report parser was re-written to more completely
and correctly parse the HID report and the HID mouse driver was
re-written to use this new parser output. These are major re-writes
and comparison with the previous code is probably not useful.